api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
PyJWT==2.10.1
httpx==0.25.0